updates for reprod. package

+++ Code/Analysis/forecast_privacy_assessment_M4_rate.R
@@ -169,4 +169,9 @@ weight_avg_prop_ident <- avg_prop_ident %>%
   summarize(weight_avg_prop = sum(num_series/sum(num_series) * avg_proportion_identified),
             total_series = sum(num_series))
 
-write.csv(weight_avg_prop_ident, paste0("../../Outputs/Results/", data_folder, "Tables/forecast_identification_probabilities.csv"), row.names=FALSE)
+if (file.exists(paste0("../../Outputs/Results/", data_folder, "Tables/"))){
+  write.csv(weight_avg_prop_ident, paste0("../../Outputs/Results/", data_folder, "Tables/forecast_identification_probabilities.csv"), row.names=FALSE)
+} else {
+  dir.create(paste0("../../Outputs/Results/", data_folder, "Tables/"))
+  write.csv(weight_avg_prop_ident, paste0("../../Outputs/Results/", data_folder, "Tables/forecast_identification_probabilities.csv"), row.names=FALSE)
+}
